+++ src/modules/sell-history/sell-history-model.ts
@@ -43,11 +43,11 @@ SellHistory.init({
     allowNull: false,
   },
   qty: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 2),
     allowNull: false,
   },
   totalPrice: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(15, 2),
     allowNull: false,
   },
   memo: {
